Move weights_and_measures to top namespace
