+++ resqpy/rq_import.py
@@ -18,7 +18,6 @@ import resqpy.olio.load_data as ld
 import resqpy.olio.write_data as wd
 import resqpy.olio.ab_toolbox as abt
 import resqpy.olio.uuid as bu
-import resqpy.olio.weights_and_measures as bwam
 import resqpy.olio.xml_et as rqet
 import resqpy.olio.vdb as vdb
 import resqpy.olio.vector_utilities as vec
@@ -31,6 +30,7 @@ import resqpy.property as rp
 import resqpy.time_series as rts
 import resqpy.surface as rqs
 import resqpy.organize as rqo
+import resqpy.weights_and_measures as bwam
 
 
 def import_nexus(resqml_file_root,                # output path and file name without .epc or .h5 extension
