+++ tests/test_weights_measures.py
@@ -4,7 +4,7 @@ import pytest
 import numpy as np
 from numpy.testing import assert_array_almost_equal
 
-import resqpy.olio.weights_and_measures as wam
+import resqpy.weights_and_measures as wam
 from resqpy.olio.exceptions import InvalidUnitError, IncompatibleUnitsError
